# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Customer Support Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List contacts
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: companyId
          in: query
          schema:
            type: number
        - name: tagId
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create a contact
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/{id}:
    get:
      operationId: getContact
      tags: [contacts]
      summary: Get contact by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    patch:
      operationId: updateContact
      tags: [contacts]
      summary: Update a contact
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete a contact
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # ─── COMPANIES ──────────────────────────────────────────────────────────────
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List companies
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create a company
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{id}:
    get:
      operationId: getCompany
      tags: [companies]
      summary: Get company by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    patch:
      operationId: updateCompany
      tags: [companies]
      summary: Update a company
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    delete:
      operationId: deleteCompany
      tags: [companies]
      summary: Delete a company
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # ─── CONVERSATIONS ──────────────────────────────────────────────────────────
...
  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List notes
      security:
        - bearerAuth: []
      parameters:
        - name: contactId
          in: query
          schema:
            type: number
        - name: conversationId
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a note
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/{id}:
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # ─── TAGS ───────────────────────────────────────────────────────────────────
...
  /deals:
    get:
      operationId: listDeals
      tags: [deals]
      summary: List deals
      security:
        - bearerAuth: []
      parameters:
        - name: pipelineId
          in: query
          schema:
            type: number
        - name: stage
          in: query
          schema:
            type: string
        - name: contactId
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of deals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deal"
    post:
      operationId: createDeal
      tags: [deals]
      summary: Create a deal
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DealInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
  /deals/{id}:
    get:
      operationId: getDeal
      tags: [deals]
      summary: Get deal by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Deal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
    patch:
      operationId: updateDeal
      tags: [deals]
      summary: Update a deal
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DealUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
    delete:
      operationId: deleteDeal
      tags: [deals]
      summary: Delete a deal
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # ─── KNOWLEDGE BASE ─────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Contact:
      type: object
      required: [id, firstName, lastName, createdAt]
      properties:
        id:
          type: number
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        companyId:
          type: ["number", "null"]
        companyName:
          type: ["string", "null"]
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        lastActivityAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    ContactInput:
      type: object
      required: [firstName, lastName]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        companyId:
          type: number
        tagIds:
          type: array
          items:
            type: number
    ContactUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        companyId:
          type: ["number", "null"]
        tagIds:
          type: array
          items:
            type: number

    # ─── COMPANY ──────────────────────────────────────────────────────────────
    Company:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        domain:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        size:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        contactCount:
          type: number
        createdAt:
          type: string
          format: date-time
    CompanyInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        domain:
          type: string
        industry:
          type: string
        size:
          type: string
    CompanyUpdate:
      type: object
      properties:
        name:
          type: string
        domain:
          type: string
        industry:
          type: string
        size:
          type: string

    # ─── CONVERSATION ─────────────────────────────────────────────────────────
...
    Note:
      type: object
      required: [id, content, authorId, createdAt]
      properties:
        id:
          type: number
        content:
          type: string
        authorId:
          type: number
        authorName:
          type: string
        contactId:
          type: ["number", "null"]
        conversationId:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
    NoteInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        contactId:
          type: number
        conversationId:
          type: number
    NoteUpdate:
      type: object
      properties:
        content:
          type: string

    # ─── TAG ──────────────────────────────────────────────────────────────────
    Tag:
      type: object
      required: [id, name, color]
      properties:
        id:
          type: number
        name:
          type: string
        color:
          type: string
...
    Deal:
      type: object
      required: [id, title, pipelineId, stage, createdAt]
      properties:
        id:
          type: number
        title:
          type: string
        pipelineId:
          type: number
        pipelineName:
          type: ["string", "null"]
        stage:
          type: string
        value:
          type: ["number", "null"]
        currency:
          type: string
        contactId:
          type: ["number", "null"]
        contactName:
          type: ["string", "null"]
        assigneeId:
          type: ["number", "null"]
        assigneeName:
          type: ["string", "null"]
        expectedCloseDate:
          type: ["string", "null"]

        probability:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
    DealInput:
      type: object
      required: [title, pipelineId, stage]
      properties:
        title:
          type: string
        pipelineId:
          type: number
        stage:
          type: string
        value:
          type: number
        currency:
          type: string
          default: USD
        contactId:
          type: number
        assigneeId:
          type: number
        expectedCloseDate:
          type: string

        probability:
          type: number
    DealUpdate:
      type: object
      properties:
        title:
          type: string
        stage:
          type: string
        value:
          type: ["number", "null"]
        contactId:
          type: ["number", "null"]
        assigneeId:
          type: ["number", "null"]
        expectedCloseDate:
          type: ["string", "null"]

        probability:
          type: ["number", "null"]

    # ─── KNOWLEDGE BASE ───────────────────────────────────────────────────────
...